Change database structure
Moved the price field to the ARTWORK table so that we can display the price just when displaying the artwork without performing any joins.
Also, by this time, no transaction has been made with the artist.
Also added the "Sold" status to Artwork so that we can quickly identify which artworks have been sold and perform further operations if needed.