[3] Setup basic database schema (#4)
* [3] Setup basic database schema

* [3] Update purchase time of artwork to be a timestamp

Timestamp avoids inconsistencies in the data. For example, if date is
used, it can happen that someone from India buys artwork on Sunday
morning. The same time is Saturday evening in America. So if absolute
dates are taken, and suppose Artist is from America, then to the
artist, the artwork is bought on a date (Sunday) which has not yet come!
Using timestamp avoids this issue.

* [3] Update as per comments on PR